+++ CHANGES.rst
@@ -11,6 +11,7 @@ History
 - Added `Terminate Application`
 - Support for strict_ssl added in `Open Application`
 - Support for arguments(dict) with `Execute Script` and `Execute Async Script`
+- Added `Tap With Positions` that uses new ActionChain of w3c
 
 1.6.4
 ----------------